Split parser core and toplevel
